refactor(spa-editor): rename weekstrip-mark to PascalCase WeekStripMark
Addresses CodeRabbit on #744: React component files under components/
must be PascalCase (cf. sibling WeekStripColumn.tsx). Rename
weekstrip-mark.tsx/.test.tsx -> WeekStripMark.tsx/.test.tsx and update the
two importers.
