+++ packages/workout-spa-editor/src/components/pages/Today/WeekStripMark.tsx
